fix account creation
ran into some problems when migrating to a new ubuntu setup. I think the
problem was async calls for role assignment and account creation, either
they were not logged in fast enough or their account was not created
fast enough for the role assignment.

now uses the profile (not for role checks, only during role assignment)