Added s to Trigger untriggered notifications
Said notifications (sent to the vendor when they have not marked a
session as completed by the end of the day) should also not have
duplicates if the session is still not cleared by the next day.

Change Vendorslots publication to only return slots from today's date
onwards. This should not be a problem as on the client, as Vendorslots
are only needed for the schedule page

Add a scheduled job to remove vendorslots older than 10 weeks (currently
not enabled) addresses #39

Add customer name to vendorsessions page